backend/requirements.txt
agno
fastapi
uvicorn
python-multipart
sqlalchemy
pydantic-settings
requests
python-dotenv
sqlite-utils
